week 6 exercise 10

+++ app/helpers/application_helper.rb
@@ -3,7 +3,7 @@ module ApplicationHelper
     return if current_user.nil?
 
     edit = link_to('Edit', url_for([:edit, item]), class: "btn btn-primary")
-    del = link_to('Destroy', item, method: :delete, form: { data: { turbo_confirm: "Are you sure ?" } }, class: "btn btn-danger")
+    del = link_to('Destroy', item, method: :delete, form: { data: { turbo_confirm: "Are you sure ?" } }, class: "btn btn-danger") if current_user.admin
     raw("#{edit} #{del}")
   end
